+++ pkg/deployment/reconcile/action_runtime_container_image_update.go
@@ -22,6 +22,7 @@ package reconcile
 
 import (
 	"context"
+	"fmt"
 	"time"
 
 	core "k8s.io/api/core/v1"
@@ -57,7 +58,7 @@ func (a actionRuntimeContainerImageUpdate) Pre(ctx context.Context) error {
 		return nil
 	}
 
-	cname, _, ok := a.getContainerDetails()
+	cname, _, _, ok := a.getContainerDetails()
 	if !ok {
 		a.log.Info("Unable to find container details")
 		return nil
@@ -104,7 +105,7 @@ func (a actionRuntimeContainerImageUpdate) Post(ctx context.Context) error {
 		}
 	}
 
-	cname, image, ok := a.getContainerDetails()
+	cname, image, _, ok := a.getContainerDetails()
 	if !ok {
 		a.log.Info("Unable to find container details")
 		return nil
@@ -155,18 +156,23 @@ func (a actionRuntimeContainerImageUpdate) Post(ctx context.Context) error {
 	})
 }
 
-func (a actionRuntimeContainerImageUpdate) getContainerDetails() (string, string, bool) {
+func (a actionRuntimeContainerImageUpdate) getContainerDetails() (string, string, string, bool) {
 	container, ok := a.action.GetParam(rotation.ContainerName)
 	if !ok {
-		return "", "", false
+		return "", "", "", false
 	}
 
 	image, ok := a.action.GetParam(rotation.ContainerImage)
 	if !ok {
-		return "", "", false
+		return "", "", "", false
 	}
 
-	return container, image, true
+	revision, ok := a.action.GetParam(rotation.ContainerRevision)
+	if !ok {
+		revision = ""
+	}
+
+	return container, image, revision, true
 }
 
 // Start starts the action for changing conditions on the provided member.
@@ -182,7 +188,7 @@ func (a actionRuntimeContainerImageUpdate) Start(ctx context.Context) (bool, err
 		return true, errors.Errorf("Client is not ready")
 	}
 
-	name, image, ok := a.getContainerDetails()
+	name, image, _, ok := a.getContainerDetails()
 	if !ok {
 		a.log.Info("Unable to find container details")
 		return true, nil
@@ -235,6 +241,14 @@ func (a actionRuntimeContainerImageUpdate) Start(ctx context.Context) (bool, err
 			// Update pod image
 			pod.Spec.Containers[id].Image = image
 
+			cstatus, ok := kresources.GetContainerStatusByName(pod, name)
+			if !ok {
+				a.log.Info("Unable to find container status")
+				return false, errors.Errorf("Unable to find container status")
+			}
+
+			a.actionImpl.action.AddParam(rotation.ContainerRevision, fmt.Sprintf("%d", cstatus.RestartCount))
+
 			if _, err := a.actionCtx.ACS().CurrentClusterCache().PodsModInterface().V1().Update(ctx, pod, meta.UpdateOptions{}); err != nil {
 				return true, err
 			}
@@ -273,7 +287,7 @@ func (a actionRuntimeContainerImageUpdate) CheckProgress(ctx context.Context) (b
 		return true, false, nil
 	}
 
-	name, image, ok := a.getContainerDetails()
+	name, image, revision, ok := a.getContainerDetails()
 	if !ok {
 		a.log.Info("Unable to find container details")
 		return true, false, nil
@@ -322,13 +336,18 @@ func (a actionRuntimeContainerImageUpdate) CheckProgress(ctx context.Context) (b
 		// Pod is still pulling image or pending for pod start
 		return false, false, nil
 	} else if s := cstatus.State.Running; s != nil {
+		a.log.Info("Checking the running")
 		// Image is changed after restart
-		if cspec.Image != cstatus.Image {
+		if cspec.Image == cstatus.Image {
 			// Image not yet updated, retry soon
-			return false, false, nil
+			return true, false, nil
 		}
 
-		return true, false, nil
+		if fmt.Sprintf("%d", cstatus.RestartCount) != revision {
+			return true, false, nil
+		}
+
+		return false, false, nil
 	} else {
 		// Unknown state
 		return false, false, nil