+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - (Feature) (Platform) Sidecar
 - (Feature) (Platform) Sidecar Initial Resources
 - (Feature) (Platform) Extend Gateway Options
+- (Feature) (Platform) Sidecar Feature
 
 ## [1.4.1](https://github.com/arangodb/kube-arangodb/tree/1.4.1) (2026-01-27)
 - (Bugfix) (Platform) Ensure Helm History limit in ArangoPlatformService